move hmm utils
